+++ integer/src/mul_ops.rs
@@ -48,6 +48,7 @@ macro_rules! impl_div_primitive_with_ibig {
 impl_div_primitive_with_ibig!(u8 u16 u32 u64 u128 usize i8 i16 i32 i64 i128 isize);
 
 impl UBig {
+    // TODO(v0.4): rename to sqr
     /// Calculate the square of the number (`x * x`).
     ///
     /// # Examples